Move class SitemapService to new XmlGenerator namespace
